fix(sast): true AST traversal for C, C++, and Rust scanners
This commit fixes the 'fake coverage' issue by replacing string-based pattern matching with genuine AST node traversal for Rust, C, and C++ quality rules. It also separates the generic dispatch logic out of the python scanner into its own quality_cgo.go file to resolve the architecture violation.
